Created File to Call seedData Function
